Moved and reworded prior things warning
The wording was unintentionally scary — in particular, it referred to “your data” which is broad and sounds really important, rather than “your prior things” which is more specific and also highly subjective as to its importance.

Also moved the warning to the Prior Things section as it really only applies to that section — not technically, but it’s the only context where it matters, as the current state is a very small amount of data. This also greatly improves the initial experience of using the app, since the first thing you see is the UI, rather than a scary warning.

Thanks to @closacco for the feedback!